analytex_microwell_designer/requirements.txt
# AnalyteX MicroWell Designer - Dependencies
# Install via: pip install -r requirements.txt

# Core CAD Engine (OpenCascade wrapper)
cadquery>=2.4.0

# GUI Framework
PyQt6>=6.5.0

# 3D Visualization
pyqtgraph>=0.13.0
PyOpenGL>=3.1.7

# Numerical Computing
numpy>=1.24.0
